#9 US2002 Fully working

+++ Java Project/base.core/src/main/java/eapli/base/agv/Repository/AGVRepository.java
@@ -3,7 +3,7 @@ package eapli.base.agv.Repository;
 import eapli.base.agv.domain.model.AGV;
 import eapli.framework.domain.repositories.DomainRepository;
 
-public interface AGVRepository extends DomainRepository<Integer, AGV> {
+public interface AGVRepository extends DomainRepository<Long, AGV> {
 
     Iterable<AGV> findAGVById(Long id);
